+++ app/models/unidom/product/product.rb
@@ -11,4 +11,10 @@ class Unidom::Product::Product < ActiveRecord::Base
   validates :measurement_unit, presence: true, length: { in: 1..self.columns_hash['measurement_unit'].limit }
   validates :packing_norm,     presence: true, length: { in: 1..self.columns_hash['packing_norm'].limit     }
 
+  has_many :source_product_associatings, class_name: 'Unidom::Product::ProductAssociating', foreign_key: :target_id #, as:     :target
+  has_many :source_products,             through:    :source_product_associatings,          source: :source
+
+  has_many :target_product_associatings, class_name: 'Unidom::Product::ProductAssociating', foreign_key: :source_id #, as:     :source
+  has_many :target_products,             through:    :target_product_associatings,          source: :target
+
 end